Handle task scheduling failures from the REST API
If RabbitMQ is unavailable and a new request is submitted,
Cachito will now fail the request and return an error to the user.

If RabbitMQ is unavailable when a request with the npm package
manager is set to the "stale" or "failed" state, the PATCH request
will still complete but a log message will state that the admin
must clean this up manually.