Verify archived git repositories
When storing a tarball for a git repository, make sure such tarball is
valid and contains a valid git repository. When invalid, delete the new
archive and fail the request.

Moreover, if a request for an already existing archive is performed,
check the existing archive. In case this archive is invalid, throw a
warning and re-create the archive.

Signed-off-by: Athos Ribeiro <athos@redhat.com>