Fix how yarn modifies lockfiles
When replacing a dependency in yarn.lock, we must completely replace the
yarn.lock entry with a new one, not just update the old one. The key for
the new entry has to be {name}@{version_in_nexus}.

Additionally, we must also replace versions of external dependencies in
all "dependencies" sections in the lockfile.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>