Do not return data from PATCH endpoint
CLOUDBLD-5734

The PATCH endpoint takes a JSON payload and updates the Cachito DB based
on that payload. One seemingly innocent thing that this endpoint does is
return the JSON representation of the request that was just updated.

One of the many barely related use cases of this endpoint is adding
dependencies to requests. We add dependencies in batches of 50. Every
time we add a dependency, it becomes part of the JSON representation of
that request. For requests with a large number of dependencies, the JSON
can quickly grow in size to the point where downloading the data takes
up the *majority* of the processing time of the entire request.

Most of the use cases of this endpoint do not utilize the returned JSON
data at all. The few that do are tied to the set_request_state() method.
Those use cases are replaced by a new get_request() method. The endpoint
no longer returns any data.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>